% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eatthis_data.R
\docType{data}
\name{eatthis_data}
\alias{eatthis_data}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame of food items:
\describe{
\item{Category}{Cateogry of food}
\item{Item}{Food item}
...
}
}
\usage{
eatthis_data
}
\description{
A dataset containing meals we have had.
}
\keyword{datasets}
